requirements.txt
requests>=2.31.0
feedparser>=6.0.11
python-dotenv>=1.0.1
pyyaml>=6.0.1
openai>=1.30.0
anthropic>=0.28.0
pre-commit>=3.7.0
ruff>=0.6.0
deep-translator>=1.11.4
